Add the possibility to limit write buffer size
Add an optional parameter to limit the write buffer size. This can be
useful on systems with limited RAM but large storage devices, where
"dense" disk images can lead to exhaustion of available memory.

If no limit is given, the tool will try to determine the amount of
free memory and use it as a limit.

Signed-off-by: Francesco Valla <valla.francesco@gmail.com>