Add options to modify checksumming behavior
If the '-w' command line option is specified, the written data is read
back and verified against the checksum contained inside the BMAP file.
This ensures at the same time that the data has been read correclty from
the image file and written with no errors.

If the '-n' command line option is specified instead, the checksum
verification is entirely skipped.

If no option is specified, the checksum is verified against data read
from the source image file.

Signed-off-by: Francesco Valla <valla.francesco@gmail.com>